requirements.txt
Flask
Flask_Login
marshmallow
python-dotenv
SQLAlchemy
Werkzeug
pandas
pillow
pymysql
requests
gunicorn
fuzzywuzzy